+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   - `Connection::send_config_update_cmd`
   - `Connection::send_internal_network_list`
   - `Connection::send_reboot_cmd`
+  - `Connection::send_shutdown_cmd`
   - `Connection::send_tor_exit_node_list`
   - `Connection::send_trusted_user_agent_list`